Move createStore to @store/helpers
